+++ backend/core/views.py
@@ -1,7 +1,7 @@
 from django.http import HttpRequest, HttpResponse
 from django.shortcuts import render
 from django.views.decorators.http import require_GET
-from jklib.dj.viewsets import ImprovedViewSet
+from django_utils_kit.viewsets import ImprovedViewSet
 from rest_framework import permissions
 from rest_framework.decorators import action
 from rest_framework.request import Request