+++ backend/django_react_starter/settings/test.py
@@ -7,6 +7,7 @@ ALLOWED_HOSTS = ["localhost", "127.0.0.1"]
 SECRET_KEY = "+qbi539m9lip0yf5t97a8n4o(_3h@3&3u30kaw@ou5ydav+s_t"
 DJANGO_SUPERUSER_EMAIL = "random-email@for-test.com"
 DJANGO_SUPERUSER_PASSWORD = "random-password-for-test"
+EMAIL_BACKEND = "django.core.mail.backends.locmem.EmailBackend"
 
 # Media
 MEDIA_ROOT = MEDIA_ROOT + "-test"