+++ backend/user/tests/test_indexers.py
@@ -1,14 +1,10 @@
-from jklib.meili.dj.test_utils import IndexerBaseTestMixin
-
 from core.tests import BaseTestCase
 from user.indexers import UserIndexer
-from user.models import User
 from user.tests.factories import UserFactory
 
 
-class UserIndexerTestCase(IndexerBaseTestMixin[User], BaseTestCase):
+class UserIndexerTestCase(BaseTestCase):
     indexer_class = UserIndexer
-    search_attribute = "indexed_name"
 
     @classmethod
     def setUpTestData(cls) -> None:
@@ -18,10 +14,13 @@ class UserIndexerTestCase(IndexerBaseTestMixin[User], BaseTestCase):
     def test_build_object(self) -> None:
         user = UserFactory()
         self.assertEqual(
-            self.indexer_class.build_object(user),
+            UserIndexer.build_object(user),
             {
                 "id": user.id,
                 "indexed_name": f"user_{user.id}",
                 "fake_type": 1,
             },
         )
+
+    def test_index_name(self) -> None:
+        self.assertEqual(UserIndexer.index_name(), "users")