+++ backend/user/views.py
@@ -1,7 +1,7 @@
 from django.contrib.auth import get_user_model, login, logout
+from django_utils_kit.permissions import IsNotAuthenticated
+from django_utils_kit.viewsets import ImprovedViewSet
 from drf_spectacular.utils import PolymorphicProxySerializer, extend_schema
-from jklib.dj.permissions import IsNotAuthenticated
-from jklib.dj.viewsets import ImprovedViewSet
 from rest_framework import status
 from rest_framework.decorators import action
 from rest_framework.permissions import IsAuthenticated